requirement.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
channels==4.0.0
charset-normalizer==3.0.1
Django==4.1.5
djangorestframework==3.14.0
idna==3.4
mysql-connector-python==8.0.32
protobuf==3.20.3
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
